fix(#382): resolve pr #390 review comments — indices type, scatter dispatch, kernel determinism
addresses unresolved coderabbit review comments on pr #390:

- cuda/hip/opencl embedding kernels: indices param type was float* with a
  numeric (int) cast that float-interpreted bit patterns and truncated to
  zero — switched to int* (cuda/hip) or as_int() (opencl) to match the
  true int32 / bit-packed-int32 upload paths in allocateintbuffer.
- opencl scatter_mean dispatch: scatter_mean_divide expects flat output
  element count but the dispatch passed row count, normalizing only the
  first row. now passes outputsize * featuresize.
- opencl gqa gradq deterministic kernel: was gated on d==0 inside a 3d
  launch, wasting (headdim-1) threads. now truly parallelizes over d,
  each work-item owns one gradquery cell and writes with =.
- vulkan scatter_mean: added missing gpudeterminism gating to route to
  scattermeandeterministicglsl, plus same flat-vs-row-count fix for the
  divide pass.
- cuda/hip crop_2d_backward, global_maxpool2d_backward deterministic
  variants: changed += to = and updated comments to reflect that each
  thread owns a unique output slot (no race, no order dependency).
- hip global_maxpool2d_backward atomic variant: corrected mislabeled
  comment (it is race-free, atomicadd was over-defensive).
- gpu determinism regression test: switched to dedicated
  aidotnetengineglobalstate collection (mutates engine.current), captures
  thread-local determinism override via blasprovider helper instead of
  merged value, disposes the directgputensorengine instance, and compares
  via bitconverter.toint32 of the raw payload bytes so +/-0 and nan
  payload differences are caught (vs assert.equal(float, float) which
  treats them as equal). uses bitconverter.getbytes/toint32 for net471
  compat (singletoint32bits is net6+).

builds clean on net10.0 + net471, zero errors.

+++ src/AiDotNet.Tensors/Engines/DirectGpu/CUDA/Kernels/CudaNeuralNetKernels.cs
@@ -1541,14 +1541,19 @@ extern ""C"" __global__ __launch_bounds__(256) void embedding_forward(
     }
 }
 
+// CUDA AllocateIntBuffer uploads true int32 (see CudaBackend.AllocateIntBuffer(int[])).
+// Declaring indices as int* (not float*) ensures correct reads; the prior signature
+// of const float* with a (int) cast would float-interpret the int's bit pattern
+// (denormal for small ints) and truncate to 0 — a long-standing bug surfaced by
+// PR #390 review (issue #382).
 extern ""C"" __global__ __launch_bounds__(256) void embedding_backward(
-    const float* gradOutput, const float* indices, float* gradEmbedding,
+    const float* gradOutput, const int* indices, float* gradEmbedding,
     int numIndices, int embeddingDim)
 {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i >= numIndices) return;
 
-    int idx = (int)indices[i];
+    int idx = indices[i];
     for (int d = 0; d < embeddingDim; d++) {
         // Atomic add for thread safety. NON-DETERMINISTIC across runs (issue #382).
         atomicAdd(&gradEmbedding[idx * embeddingDim + d], gradOutput[i * embeddingDim + d]);
@@ -1560,7 +1565,7 @@ extern ""C"" __global__ __launch_bounds__(256) void embedding_backward(
 // accumulates contributions where indices[i] == v. No atomics; accumulation order is
 // identical across runs.
 extern ""C"" __global__ __launch_bounds__(256) void embedding_backward_deterministic(
-    const float* gradOutput, const float* indices, float* gradEmbedding,
+    const float* gradOutput, const int* indices, float* gradEmbedding,
     int numIndices, int embeddingDim, int vocabSize)
 {
     int v = blockIdx.y * blockDim.y + threadIdx.y;
@@ -1569,7 +1574,7 @@ extern ""C"" __global__ __launch_bounds__(256) void embedding_backward_determini
 
     float sum = 0.0f;
     for (int i = 0; i < numIndices; i++) {
-        if ((int)indices[i] == v) {
+        if (indices[i] == v) {
             sum += gradOutput[i * embeddingDim + d];
         }
     }
